refactor startup
